Allow to configure atracdenc to use float instead of double
The default is double for hystorical reason.
Actually no any reason to use double precision float point
calculation for all processing. AT3P encoding will use float
except places were we really need double precision.
So this default will be changed to float in near future and
this option will be removed.

+++ src/atrac3denc.cpp
@@ -130,7 +130,7 @@ TAtrac3MDCT::TGainModulatorArray TAtrac3MDCT::MakeGainModulatorArray(const TAtra
 
 TFloat TAtrac3Encoder::LimitRel(TFloat x)
 {
-    return std::min(std::max(x, GainLevel[15]), GainLevel[0]);
+    return std::min(std::max((double)x, GainLevel[15]), GainLevel[0]);
 }
 
 void TAtrac3Encoder::ResetTransientParamsHistory(int channel, int band)
